requirements.txt
Django==6.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
redis==5.0.1